third_party/blink/web_tests/external/wpt/html/syntax/serializing-html-fragments/serializing-expected.txt

This is a testharness.js-based test.
Found 6 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] innerHTML 7 <a b="<"></a>
  assert_equals: expected "<a b=\\"<\\"></a>" but got "<a b=\\"&lt;\\"></a>"
[FAIL] innerHTML 8 <a b=">"></a>
  assert_equals: expected "<a b=\\">\\"></a>" but got "<a b=\\"&gt;\\"></a>"
[FAIL] innerHTML 9 <a href="javascript:&quot;<>&quot;"></a>
  assert_equals: expected "<a href=\\"javascript:&quot;<>&quot;\\"></a>" but got "<a href=\\"javascript:&quot;&lt;&gt;&quot;\\"></a>"
[FAIL] outerHTML 7 <span><a b="<"></a></span>
  assert_equals: expected "<span><a b=\\"<\\"></a></span>" but got "<span><a b=\\"&lt;\\"></a></span>"
[FAIL] outerHTML 8 <span><a b=">"></a></span>
  assert_equals: expected "<span><a b=\\">\\"></a></span>" but got "<span><a b=\\"&gt;\\"></a></span>"
[FAIL] outerHTML 9 <span><a href="javascript:&quot;<>&quot;"></a></span>
  assert_equals: expected "<span><a href=\\"javascript:&quot;<>&quot;\\"></a></span>" but got "<span><a href=\\"javascript:&quot;&lt;&gt;&quot;\\"></a></span>"
Harness: the test ran to completion.
